Tour information for concerts in the spRingsteen dataset.

Data describing the tours associated with concerts played by Bruce Springsteen both solo and with numerous bands from the years 1973 to present day. Note that concerts prior to 1973 and non-tour, e.g., practice shows, promotion shows, have been removed. Furthermore some of the shows are associated with more than one tour as such some of the entries from concerts appear twice. Can be joined with setlists or concerts using gig_key.

Usage

tours

Format

A data frame with 2 variables:

gig_key

Primary key of the data frame.

tour

Tour associated with the concert. Note some concerts have more than one tour associated with them.

Source

http://brucebase.wikidot.com/

Examples

library(dplyr)
# How many shows were on each tour?

tours %>%
  count(tour, sort = TRUE)
#> # A tibble: 24 × 2
#>    tour                                                   n
#>    <chr>                                              <int>
#>  1 Non-tour Shows                                       575
#>  2 Springsteen On Broadway                              268
#>  3 The River Tour                                       213
#>  4 The Wild, The Innocent & The E Street Shuffle Tour   197
#>  5 Born In The U.S.A. Tour                              156
#>  6 Greetings From Asbury Park Tour                      147
#>  7 Wrecking Ball Tour                                   134
#>  8 The Reunion Tour                                     132
#>  9 The Ghost Of Tom Joad Tour                           128
#> 10 The Rising Tour                                      120
#> # … with 14 more rows
#> # ℹ Use `print(n = ...)` to see more rows